+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Removed deprecated `AnalysisOptions.from` use `AnalysisOptions.fromMap` instead
 - Removed deprecated `Config.linesOfCodeWarningLevel` use `Config.linesOfExecutableCodeWarningLevel` instead
+- Removed deprecated `MetricsAnalysisRecorder.startRecordFile` and `MetricsAnalysisRecorder.endRecordFile` use `MetricsRecordsStore.recordFile` instead
 - Add static code anti-patterns long-parameter-list
 - Add static code diagnostic potential-null-dereference
 - Set min SDK version to >=2.8.0.