+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Removed deprecated `AnalysisOptions.from` use `AnalysisOptions.fromMap` instead
 - Removed deprecated `Config.linesOfCodeWarningLevel` use `Config.linesOfExecutableCodeWarningLevel` instead
 - Removed deprecated `MetricsAnalysisRecorder.startRecordFile` and `MetricsAnalysisRecorder.endRecordFile` use `MetricsRecordsStore.recordFile` instead
+- **Breaking Change:** `MetricsAnalyzer.runAnalysis` now accept array with folder paths
 - Add static code anti-patterns long-parameter-list
 - Add static code diagnostic no-equal-arguments
 - Set min SDK version to >=2.8.0.