+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-# Unreleased (2.0.0)
+# 2.0.0
 
 - Removed deprecated `AnalysisOptions.from` use `AnalysisOptions.fromMap` instead
 - Removed deprecated `Config.linesOfCodeWarningLevel` use `Config.linesOfExecutableCodeWarningLevel` instead